shiksha-api/app-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure dependencies
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# OpenAI and LLM dependencies
litellm
autogen-core

# Database and storage
asyncpg==0.29.0
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Environment and configuration
python-dotenv==1.0.0
